requirements.txt
pytest
numpy
soundfile
pillow
